restructure temporary package
