+++ claimer/claimer/errors.go
@@ -1,4 +1,4 @@
-package renamethis
+package claimer
 
 import (
 	"fmt"
