+++ claimer/claimer/workers.go
@@ -1,131 +1,25 @@
-package renamethis
+package claimer
 
 import (
-	"context"
 	"encoding/hex"
 	"errors"
 	"fmt"
 	"math/rand"
-	"os"
 	"strings"
 	"time"
 
-	brpc "github.com/kava-labs/binance-chain-go-sdk/client/rpc"
-	btypes "github.com/kava-labs/binance-chain-go-sdk/common/types"
-	bkeys "github.com/kava-labs/binance-chain-go-sdk/keys"
-	log "github.com/sirupsen/logrus"
-	amino "github.com/tendermint/go-amino"
-	"golang.org/x/sync/semaphore"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/kava-labs/go-sdk/keys"
-	kava "github.com/kava-labs/kava/app"
-	bep3 "github.com/kava-labs/kava/x/bep3/types"
-	tmlog "github.com/tendermint/tendermint/libs/log"
-	rpcclient "github.com/tendermint/tendermint/rpc/client/http"
-	tmtypes "github.com/tendermint/tendermint/types"
-
+	brpc "github.com/kava-labs/binance-chain-go-sdk/client/rpc"
+	btypes "github.com/kava-labs/binance-chain-go-sdk/common/types"
 	"github.com/kava-labs/go-tools/claimer/config"
 	"github.com/kava-labs/go-tools/claimer/server"
+	bep3 "github.com/kava-labs/kava/x/bep3/types"
+	log "github.com/sirupsen/logrus"
+	amino "github.com/tendermint/go-amino"
+	rpcclient "github.com/tendermint/tendermint/rpc/client/http"
+	tmtypes "github.com/tendermint/tendermint/types"
 )
-
-var JobQueueSize = 1000
-
-// KavaClaimer is a worker that sends claim transactions on Kava
-type KavaClaimer struct {
-	Keybase keys.KeyManager
-	Status  bool
-}
-
-func Main(ctx context.Context, c config.Config) {
-	// Load kava claimers
-	sdkConfig := sdk.GetConfig()
-	kava.SetBech32AddressPrefixes(sdkConfig)
-	cdc := kava.MakeCodec()
-
-	// SETUP CLAIMERS --------------------------
-	var kavaClaimers []KavaClaimer
-	for _, kavaMnemonic := range c.Kava.Mnemonics {
-		kavaClaimer := KavaClaimer{}
-		keyManager, err := keys.NewMnemonicKeyManager(kavaMnemonic, kava.Bip44CoinType)
-		if err != nil {
-			log.Error(err)
-		}
-		kavaClaimer.Keybase = keyManager
-		kavaClaimer.Status = true
-		kavaClaimers = append(kavaClaimers, kavaClaimer)
-	}
-
-	// SETUP KAVA CLIENT --------------------------
-	// Start Kava HTTP client
-	http, err := rpcclient.New(c.Kava.Endpoint, "/websocket")
-	if err != nil {
-		panic(err)
-	}
-	http.Logger = tmlog.NewTMLogger(tmlog.NewSyncWriter(os.Stdout))
-	err = http.Start()
-	if err != nil {
-		panic(err)
-	}
-
-	// SETUP BNB CLIENT --------------------------
-	// Set up Binance Chain client
-	bncNetwork := btypes.TestNetwork
-	if c.BinanceChain.ChainID == "Binance-Chain-Tigris" {
-		bncNetwork = btypes.ProdNetwork
-	}
-	bnbClient := brpc.NewRPCClient(c.BinanceChain.Endpoint, bncNetwork)
-	bnbKeyManager, err := bkeys.NewMnemonicKeyManager(c.BinanceChain.Mnemonic)
-	if err != nil {
-		panic(err)
-	}
-	bnbClient.SetKeyManager(bnbKeyManager)
-
-	log.Info("Starting server...")
-	claims := make(chan server.ClaimJob, JobQueueSize)
-	s := server.NewServer(claims)
-	go s.StartServer()
-
-	sem := semaphore.NewWeighted(int64(len(kavaClaimers)))
-
-	// RUN WORKERS --------------------------
-	for {
-		select {
-		case <-ctx.Done():
-			err = s.Shutdown(ctx)
-			if err != nil {
-				log.Error(err)
-			}
-			return
-		case claim := <-claims:
-			switch strings.ToUpper(claim.TargetChain) {
-			case server.TargetKava:
-				if err := sem.Acquire(ctx, 1); err != nil {
-					log.Error(err)
-					return
-				}
-
-				go func() {
-					defer sem.Release(1)
-					Retry(10, 20*time.Second, func() (err ClaimError) {
-						err = claimOnKava(c.Kava, http, claim, cdc, kavaClaimers)
-						return
-					})
-				}()
-				break
-			case server.TargetBinance, server.TargetBinanceChain:
-				go func() {
-					Retry(10, 15*time.Second, func() (err ClaimError) {
-						err = claimOnBinanceChain(bnbClient, claim)
-						return
-					})
-				}()
-				break
-			}
-		}
-	}
-}
 
 func claimOnBinanceChain(bnbHTTP brpc.Client, claim server.ClaimJob) ClaimError {
 	swapID, err := hex.DecodeString(claim.SwapID)
